test: add tests against preview url
